car_aero_gnn/requirements.txt
# Core dependencies
torch>=2.0.0
torch_geometric>=2.3.0
torch_scatter>=2.1.0
torch_sparse>=0.6.0

# Numerical computing
numpy>=1.24.0
scipy>=1.10.0

# Visualization
pyvista>=0.40.0
vtk>=9.2.0
matplotlib>=3.7.0
imageio>=2.30.0
imageio-ffmpeg>=0.4.0

# Data handling
pyyaml>=6.0
tqdm>=4.65.0
h5py>=3.8.0

# Experiment tracking (optional)
wandb>=0.15.0

# Development tools (optional)
pytest>=7.3.0
black>=23.0.0
flake8>=6.0.0

# GPU acceleration (optional, uncomment if needed)
# tensorboard>=2.13.0
# torchvision>=0.15.0
